i386: build the full shared root filesystem
Extract the architecture-independent full-rootfs profile for MIPS and i386, install the complete GCC userland and development files into a deterministic 16 MiB UFS image, and keep PCC confined to MIPS targets. Extend the i386 mapping/link boundary and libc ABI support required by the larger embedded image, and make QEMU exercise representative full-userland commands. N64/VR4300 and Ci20 full builds with native PCC pass after the shared-profile conversion.

+++ sys/kernel/kern_sysctl.c
@@ -162,7 +162,7 @@ __sysctl()
 {
     register struct sysctl_args *uap = (struct sysctl_args*) u.u_arg;
     int error;
-    u_int oldlen = 0;
+    size_t oldlen = 0;
     sysctlfn *fn;
     int name [CTL_MAXNAME];
